+++ spec/www/spec/db-tx-string-test.js
@@ -593,7 +593,7 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'string parameter value manipulation test with UTF-8 2-octet character é', function(done) {
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
+          // XXX BUG FIXED ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("UTF8-2-octet-upper-value-string-test.db", "1.0", "Demo", DEFAULT_SIZE);
 
@@ -650,7 +650,7 @@ var mytests = function() {
 
         it(suiteName + 'string parameter value manipulation test with UTF-8 3-octet character €', function(done) {
           if (isWP8) pending('SKIP for WP(8)');
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
+          // XXX BUG FIXED ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("UTF8-3-octet-string-upper-value-test.db", "1.0", "Demo", DEFAULT_SIZE);
 
@@ -1004,7 +1004,7 @@ var mytests = function() {
 
         it(suiteName + 'Inline string manipulation test with UTF-8 2/3 octet characters', function(done) {
           if (isWP8) pending('SKIP for WP(8)');
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
+          // XXX BUG FIXED ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("Inline-UTF8-string-manipulation-test.db", "1.0", "Demo", DEFAULT_SIZE);
 
@@ -1033,7 +1033,7 @@ var mytests = function() {
 
         it(suiteName + 'string parameter manipulation test with UTF-8 2/3 octet characters', function(done) {
           if (isWP8) pending('SKIP for WP(8)');
-          if (isAndroid && !isWebSql && !isImpl2) pending('XXX SKIP: BUG on Android (default Android-sqlite-evcore-native-driver access implementation)'); // XXX ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
+          // XXX BUG FIXED ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#19
 
           var db = openDatabase("UTF8-string-parameter-manipulation-test.db", "1.0", "Demo", DEFAULT_SIZE);